Move gtm file
